+++ lib/load.php
@@ -111,6 +111,8 @@ require __DIR__ . '/compat/wordpress-6.5/kses.php';
 require __DIR__ . '/compat/wordpress-6.5/interactivity-api/class-wp-interactivity-api.php';
 require __DIR__ . '/compat/wordpress-6.5/interactivity-api/class-wp-interactivity-api-directives-processor.php';
 require __DIR__ . '/compat/wordpress-6.5/interactivity-api/interactivity-api.php';
+require __DIR__ . '/compat/wordpress-6.5/class-wp-script-modules.php';
+require __DIR__ . '/compat/wordpress-6.5/scripts-modules.php';
 
 // Experimental features.
 require __DIR__ . '/experimental/block-editor-settings-mobile.php';
@@ -119,12 +121,11 @@ require __DIR__ . '/experimental/navigation-theme-opt-in.php';
 require __DIR__ . '/experimental/kses.php';
 require __DIR__ . '/experimental/l10n.php';
 require __DIR__ . '/experimental/synchronization.php';
+require __DIR__ . '/experimental/script-modules.php';
 
 if ( gutenberg_is_experiment_enabled( 'gutenberg-no-tinymce' ) ) {
 	require __DIR__ . '/experimental/disable-tinymce.php';
 }
-
-require __DIR__ . '/experimental/modules/class-gutenberg-modules.php';
 
 // Fonts API / Font Face.
 remove_action( 'plugins_loaded', '_wp_theme_json_webfonts_handler' ); // Turns off WordPress 6.0's stopgap handler.
@@ -135,8 +136,8 @@ require __DIR__ . '/experimental/fonts/font-library/class-wp-font-library.php';
 require __DIR__ . '/experimental/fonts/font-library/class-wp-font-family-utils.php';
 require __DIR__ . '/experimental/fonts/font-library/class-wp-font-family.php';
 require __DIR__ . '/experimental/fonts/font-library/class-wp-rest-font-families-controller.php';
+require __DIR__ . '/experimental/fonts/font-library/class-wp-rest-font-faces-controller.php';
 require __DIR__ . '/experimental/fonts/font-library/class-wp-rest-font-collections-controller.php';
-require __DIR__ . '/experimental/fonts/font-library/class-wp-rest-autosave-font-families-controller.php';
 require __DIR__ . '/experimental/fonts/font-library/font-library.php';
 
 // Load the Font Face and Font Face Resolver, if not already loaded by WordPress Core.