+++ test/e2e/specs/site-editor/user-global-styles-revisions.spec.js
@@ -14,7 +14,7 @@ test.use( {
 	},
 } );
 
-test.describe( 'Global styles revisions', () => {
+test.describe( 'Style Revisions', () => {
 	let stylesPostId;
 	test.beforeAll( async ( { requestUtils } ) => {
 		await Promise.all( [